requirements.txt
anthropic>=0.40.0
rich>=13.7.0
pyserial>=3.5
python-dotenv>=1.0.0
colorama>=0.4.6
